// src/app/models/sjcs/AplicacionRetencionItem.ts
export class AplicacionRetencionItem {
    anioMes: string;
    tipoRetencion: string;
    importeAntAplicaRetencion: string;
    importeAntRetenido: string;
    importeAplicaRetencion: string;
    importeRetenido: string;
    importeTotAplicaRetencion: string;
    importeTotRetenido: string;
    importeSmi: string;
    colegiado: string;
    nombrePago: string;
}