// src/app/models/sjcs/RetencionesAplicadasItem.ts
export class RetencionesAplicadasItem {
    id: string;
    tipoRetencion: string;
    numColegiado: string;
    nombre: string;
    idDestinatario: string;
    nombreDestinatario: string;
    desDestinatario: string;
    fechaInicio: Date;
    fechaFin: Date;
    fechaRetencion: Date;
    importeRetenido: string;
    idPagosjg: string;
    anioMes: string;
    pagoRelacionado: string;
    abonoRelacionado: string;
    idCobro: string;
    idRetencion: string;
    idInstitucion: string;
    idPersona: string;
    fechaDesde: string;
    fechaHasta: string;
    importePago: string;
}